Add interface for the communication
This change introduces a basic interface for the chat channel. In fact,
it is needed to be able to test the functionality of the socket upon
sending and receiving.

This changes introduces the basic exchange to emit and to listen to
messages based on events thanks to Socket.io. However, this last aspect
needs to be tested to ensure it's functionality.