Perform the communication both ways sending and receiving
This change parses the chat data received from the server and shows a
Toast message to the user. A Json is sent to the server using the
library Gson. However, the parse is done manually.

The UI is still to be completed ina different change to show the
information directly in the UI.