Use an array list to recognize selected tools
Before this change, the canvas went across all the children nodes to
know if an element is selected. This is fine for a small drawing but
given a large drawing, it can be problematic.

Instead of iterating across all children an array that defines the
selected items is being used.Please enter the commit message for your
changes. Lines starting