Add condition if the userId is the drawingOwner allow him to modify
Before this change, there was a race condition between receiving the
users first and receiving the gallery first. This caused the user to not
be able to occasionally not be able to modify or delete his drawings.

To correct this race condition, this change adds another condition so it
can validate the user as owner and it is to check if the ownerId of a
drawing is equal to the userId.