Set the gallery after receiving the lists of users
Sometimes the creators name appears on a drawing in the gallery and
sometimes it doesn't. This happens because of a race condition. The HTTP
request for the gallery can be received before or after receiving the
names list. This happens to be because of the asynchronise nature of the
socket and http requests.