Remove the intent and the action associated to the notification
Before this change, the notification had a redirection to the landing
page. This can have some problems because we cannot leave the team and
drawing from the landing page. Also, it will be weird not seeing the
teams on the landing page especially that we dont land on the last
activity.